memory optimisation
There is no need to save the whole lineage for a given commit

Just remember the parent from each repository, and compare
those lists, since a commit can really stand in for all its
lineage in its origin repository.

This should be a huge memory saver when processing repositories
with a lot of commits.