mark the original source for annotated tags too
URL: https://rt.cpan.org/Ticket/Display.html?id=94816
Reporter: marc.towersap@gmail.com
Subject: git-stitch-repo on unix seemed to miss labels for a particular repo

I am (on centos 6.5) invoking git-stitch-repo.  It seemed to work ok other
than I noticed some label oddities.  I imported 6 repositories in one shot,
and it seemed to mess up the tags.

For example, in the original repo "A", I do git tag -nl, I see a single
tag, which is an annotated tag called v6.3.3.2.

Post-git-stitch-repo, I see
v6.3.3.2 with the correct message
v6.3.3.2-A with the commit message.

but that seems to have thrown off things.  I see extra tags that don't have
a -[letter],

I think it's annotated tags that it's puking on.
It creates two tags for each annotated tag
the original tag with the original message (without -[letter]), and a
second tag with the -[letter] and the commit message it's attached to.