Git::FastExport does not run `git fast-export` by itself any more
**** INCOMPATIBLE CHANGE ****

This commit changes the interface for Git::FastExport->new().

The constructor now takes a single parameter which is a filehandle
connected to a `git fast-export` stream. This makes it easier to handle
any source of fast-import streams.

Git::FastExport::Stitch has also been modified to stay compatible with
this change, and its stitch() method now expects either a Git::Repository
object or a repository directory.

The test scripts have been modified accordingly.