Do not modify the directory name when stitching repositories
The source repositories names are used to produce branch names in the
generated repository. They are also used to produce top-level directory
names in the new directory structure, if none is provided.

This patch picks the directory name before applying the transformations
to make it suitable as a branch name.  (The only transformation fone is
the removal of the .git extension, if any.)

Fixes RT#129540.